在Reader中增加额外载波的逻辑

+++ include/gnuradio/reader/gate.h
@@ -32,7 +32,7 @@ public:
      * class. reader::gate::make is the public interface for
      * creating new instances.
      */
-    static sptr make(int sample_rate);
+    static sptr make(float sample_rate);
 };
 
 } // namespace reader
